Move _CreateSource to its own package to make it pickelable
This is needed because executing a Read as a
Impulse() | ParDo(Split) | ParDo(Read) requires the source to be
pickelable in the splits we emit in ParDo(Split).
